share validation logic between generators
Also, to make the sharing easier, remove the weird restriction that
the pacman generator can only resolve "group:" and "except:" relations
on Requires, but not on Provides/Conflicts/Replaces.

With this, #4 is almost done, but I want to cleanup the one TODO in the
debian generator before actually closing it.