support prerelease packages (alpha and beta)
The "PrereleaseType" and "PrereleaseNo" config values are incorporated
into version strings for different package formats in a way that makes
the package managers detect upgrades and downgrades correctly. Tested
with the sequence:

  1.0.0-alpha.1 < 1.0.0-alpha.3 < 1.0.0-beta.2 < 1.0.0-beta.11 < 1.0.0

For Debian and RPM the "prerelease" part of version is separated
from the main part with "~". It's an established convention on Debian.

Fedora has other recommendations, which, unfortunately, omit this
information from the version string. The Debian convention still works
with RPM and it's clearer to the end-user.

Pacman doesn't allow "~" in version strings. It allows "_", but using
"_" results in the "1.0.0-beta.11 -> 1.0.0" transition being
recognized as a downgrade. Dropping the separator altogether fixes the
issue.