change the way how alpha/beta versions are represented in the pkg.toml
The format with version, prerelease type and prerelease version works
well for the API of libpackagebuild, but I found it unnecessarily
cumbersome for the user-exposed format of pkg.toml. Compared to his
original proposal,

    version = "1.2.3"
    prereleaseType = "alpha"
    prereleaseNo = 4

the new format,

    version = "1.2.3"
    alpha = 4

reads much more pleasantly.

With this, basically nothing from Jakub's original version of this
feature is left over in holo-build, but I left his commit in this branch
to give proper attribution for the original idea.